# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shivam Animes Premium Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a premium user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [admin]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{id}/reset-device:
    post:
      operationId: resetUserDevice
      tags: [admin]
      summary: Reset user device lock
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Device reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{id}/force-logout:
    post:
      operationId: forceLogoutUser
      tags: [admin]
      summary: Force logout user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/generate:
    post:
      operationId: generateUser
      tags: [admin]
      summary: Generate random user credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateUserInput"
      responses:
        "201":
          description: Generated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Anime
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    User:
      type: object
      required: [id, username, role, isPremium, isActive, createdAt]
      properties:
        id: { type: string }
        username: { type: string }
        role: { type: string }
        isPremium: { type: boolean }
        premiumExpiresAt: { type: ["string", "null"] }
        premiumDays: { type: ["integer", "null"] }
        isActive: { type: boolean }
        isSuspended: { type: boolean }
        deviceId: { type: ["string", "null"] }
        lastIp: { type: ["string", "null"] }
        lastCountry: { type: ["string", "null"] }
        lastActiveAt: { type: ["string", "null"] }
        loginCount: { type: integer }
        createdAt: { type: string }
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    UserInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        isPremium: { type: boolean }
        premiumDays: { type: integer }
    UserUpdate:
      type: object
      properties:
        isPremium: { type: boolean }
        premiumDays: { type: integer }
        premiumExpiresAt: { type: ["string", "null"] }
        isSuspended: { type: boolean }
        isActive: { type: boolean }
    GenerateUserInput:
      type: object
      required: [premiumDays]
      properties:
        username: { type: ["string", "null"] }
        password: { type: ["string", "null"] }
        premiumDays: { type: integer }
...